# api/swagger.yml
openapi: 3.0.3
info:
  title: 'Webapp swagger - OpenAPI 3.0'
  description: "This is a Webapp swagger"
  version: 0.0.1 (v3)
servers:
  - url: https://staging.jacekplaymaker.usermd.net/api/v3/
  - url: https://dev.jacekplaymaker.usermd.net/api/v3/
  - url: http://localhost:{port}/api/v3/
    variables:
      port:
        default: '8000'
paths:
...
  /transfers/transfer-status/:
    patch:
      operationId: update_profile_transfer_status
      description: "Updates profile transfer status"
      tags:
        - transfer status
      security:
        - jwtAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/X-Language'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                contact_email:
                  type: string
                  description: "Profile contact email"
                phone_number:
                  type: object
                  properties:
                    dial_code:
                      type: integer
                      description: "Represents a phone number dial code"
                    number:
                      type: string
                      description: "Profile phone number"
                status:
                  type: integer
                  description: "Id of transfer status object. Check avaiable objects at /transfers/list-transfer-status/"
                additional_info:
                  type: array
                  items:
                    type: integer
                    description: "Id of an object from: /transfers/transfer-status/additional-info/"
                benefits:
                  type: array
                  items:
                    type: integer
                    description: "Id of an object from: /transfers/transfer-request/benefits"
                salary:
                  type: integer
                  description: "Id of an object from: /transfers/transfer-request/salary"
                number_of_trainings:
                  type: integer
                  description: "Id of an object from: /transfers/transfer-request/number-of-trainings/"
                league:
                  type: array
                  items:
                    type: integer
                    description: "Id of an highest parent league taken from: /clubs/leagues/highest-parents/"
                is_anonymous:
                  type: boolean
                  description: "If true, the transfer status will be anonymous and not linked to a specific team contributor."
            example:
              {
                "benefits": [ 1,2 ],
                "status": 1,
                "gender": "M",
                "salary": 1,
                "contact_email": "email@email.com",
                "phone_number": { "dial_code": 1, "number": 123 },
                "additional_info": [ 1, 2 ],
                "number_of_trainings": 1,
                "is_anonymous": True,
              }
      responses:
        "200":
          description: "Status 200 when update succeded"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferStatusGet'
        "403":
          description: "User is not allowed to do this action"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"
        "404":
          description: "Transfer status for given profile not found."
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"
    post:
      operationId: profile_transfer_status_create
      description: "Create profile transfer status object"
      security:
        - jwtAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/X-Language'
      tags:
        - transfer status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contact_email:
                  type: string
                  description: "Profile contact email"
                phone_number:
                  type: object
                  properties:
                    dial_code:
                      type: integer
                      description: "Represents a phone number dial code"
                    number:
                      type: string
                      description: "Profile phone number"
                status:
                  type: integer
                  description: "Id of transfer status object. Check avaiable objects at /transfers/list-transfer-status/"
                additional_info:
                  type: array
                  items:
                    type: integer
                    description: "Id of an object from: /transfers/transfer-status/additional-info/"
                benefits:
                  type: array
                  items:
                    type: integer
                    description: "Id of an object from: /transfers/transfer-requests/benefits/"
                salary:
                  type: integer
                  description: "Id of transfer status object. Check avaiable objects at /transfers/transfer-request/salary/"
                number_of_trainings:
                  type: integer
                  description: "Id of transfer status object. Check avaiable objects at /transfers/transfer-request/number-of-trainings/"
                league:
                  type: array
                  items:
                    type: integer
                    description: "Id of an highest parent league taken from: /clubs/leagues/highest-parents/"
                is_anonymous:
                  type: boolean
                  description: "If true, the transfer status will be anonymous and not linked to a specific team contributor."
              required: [ "status", "additional_info", "league" ]
      responses:
        "201":
          description: "Returns created profile transfer status object."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferStatusGet'
        "403":
          description: "User is not allowed to do this action"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"

    delete:
      operationId: delete_profile_transfer_status
      description: "Deletes profile transfer status"
      tags:
        - transfer status
      security:
        - jwtAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/X-Language'
      responses:
        "204":
          description: "Object deleted successful"
        "403":
          description: "User is not allowed to do this action"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"
        "404":
          description: "Transfer status for given profile not found."
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"
...
  /transfers/transfer-request/:
    patch:
      operationId: update_profile_transfer_request
      description: "Updates profile transfer request"
      tags:
        - transfer requests
      security:
        - jwtAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/X-Language'
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequestUpdate"
      responses:
        "200":
          description: "Status 200 when update succeded"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferStatusGet'
        "403":
          description: "User is not allowed to do this action"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"
        "404":
          description: "Transfer status for given profile not found."
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"
    post:
      operationId: profile_transfer_request_create
      description: "Create profile transfer request object"
      security:
        - jwtAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/X-Language'
      tags:
        - transfer requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequestCreate"
      responses:
        "201":
          description: "Returns created profile transfer status object."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferStatusGet'
        "403":
          description: "User is not allowed to do this action"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"

    delete:
      operationId: delete_profile_transfer_request
      description: "Deletes profile transfer request"
      tags:
        - transfer requests
      security:
        - jwtAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/X-Language'
      responses:
        "204":
          description: "Object deleted successful"
        "403":
          description: "User is not allowed to do this action"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"
        "404":
          description: "Transfer request for given profile not found."
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    description: "Represents bool as string"
                  detail:
                    type: string
                    description: "Failure detail message"
...
components:
  schemas:
...
    TransferStatusGet:
      type: object
      properties:
        contact_email:
          type: string
          description: "Contact email address."
        phone_number:
          type: object
          properties:
            dial_code:
              type: string
              description: "Represents dial code for phone number."
            number:
              type: string
              description: "Contact mobile phone number."
          required: [ "dial_code", "number" ]
        status:
          type: object
          properties:
            id:
              type: string
              description: "Transfer status id."
            name:
              type: string
              description: "Transfer status name"
          required: [ "id", "name" ]
        additional_info:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: "Object id."
              name:
                type: string
                description: "Object name"
            required: [ "id", "name" ]
        benefits:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                description: "Object id."
              name:
                type: string
                description: "Object name"
            required: [ "id", "name" ]
        number_of_trainings:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
            required: [ "id", "name" ]
        salary:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
          required: [ "id", "name" ]
        league:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                description: "Object id."
              name:
                type: string
                description: "Object name"
            required: [ "id", "name" ]
        is_anonymous:
          type: boolean
          description: "If true, the transfer status will be anonymous and not linked to a specific team contributor."
      required: [ "contact_email", "phone_number", "status", "additional_info", "league", "benefits", "salary", "number_of_trainings" ]
      example:
        {
          "contact_email": "email@email.com",
          "phone_number": {
            "dial_code": "+1",
            "number": 123
          },
          "status": {
            "id": "1",
            "name": "Szukam klubu"
          },
          "additional_info": [
            {
              "id": "1",
              "name": "Dostępny od zaraz"
            },
            {
              "id": "2",
              "name": "Wracam po kontuzji"
            }
          ],
          "league": [
            {
              "id": 13,
              "name": "B klasa",
              "is_parent": true
            }
          ],
          "benefits": [
            {
              "id": "1",
              "name": "Zwrot lub organizacja transportu"
            },
            {
              "id": "2",
              "name": "Zwrot za buty lub rękawice"
            }
          ],
          "salary": {
            "id": "1",
            "name": "0 - 499 zł"
          },
          "number_of_trainings": [
            {
              "id": "1",
              "name": "1-2 treningi tygodniowo"
            }],
          "is_anonymous": True,
        }
    TransferRequestUpdate:
      type: object
      properties:
        requesting_team:
          type: integer
          description: "Team contributor id."
        gender:
          type: string
        status:
          type: integer
          description: "Status object id."
        position:
          type: array
          items:
            type: integer
            description: "Position object id's"
        number_of_trainings:
          type: integer
          description: "Object id"
        benefits:
          type: array
          items:
            type: integer
            description: "Additional information object id's"
        salary:
          type: integer
          description: "Object id"
        contact_email:
          type: string
        phone_number:
          type: object
          properties:
            dial_code:
              type: integer
            number:
              type: string
        is_anonymous:
          type: boolean
          description: "If true, the transfer request will be anonymous and not linked to a specific team contributor."
    TransferRequestCreate:
      type: object
      properties:
        requesting_team:
          type: integer
          description: "Team contributor id."
        gender:
          type: string
        status:
          type: integer
          description: "Status object id."
        position:
          type: array
          items:
            type: integer
            description: "Position object id's"
        number_of_trainings:
          type: integer
          description: "Object id"
        benefits:
          type: array
          items:
            type: integer
            description: "Additional information object id's"
        salary:
          type: integer
          description: "Object id"
        contact_email:
          type: string
        phone_number:
          type: object
          properties:
            dial_code:
              type: integer
            number:
              type: string
        is_anonymous:
          type: boolean
          description: "If true, the transfer request will be anonymous and not linked to a specific team contributor."
      required: [ "requesting_team", "status", "player_position" ]
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT